% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToCentral.R
\name{GetEmailToken}
\alias{GetEmailToken}
\title{Retrieve ODK Central API Token}
\usage{
GetEmailToken(central_url, central_email, central_password)
}
\arguments{
\item{central_url}{The url which pounts to an ODK Central sever}

\item{central_email}{The email linked to your odk central account}

\item{central_password}{The password used for your central account}
}
\value{
This function will return a token. This token is useful for
}
\description{
To connect to ODK Central users must retrieve a token. To use this function, you will need:
1. An ODK Central server (Here are instructions to
set up ODK Central: https://docs.getodk.org/central-setup/)
2. An account on the server (you will have to register with an email and a password)
}
\details{
In order to keep your credential safe, it is recommended that you save your username and
password as "environment variables".
}
\examples{
\dontrun{
my_central_url <- "https://myrandomurl.com"
my_central_email <- Sys.getenv("odk_username")
my_central_password <- Sys.getenv("odk_password")

token<-GetEmailToken(central_url=my_central_url,
                     central_email = my_central_email,
                     central_password = my_central_password)
}
}
